new Assets
Added floor mat, creation box. Added some extra industrial materials to the material pack incase you would like to use any. But will delete later in interest of saving space if not used for something.. Visual Studio subscription not working so not updates to script in unity but i have some files i would like to add and try saved on the computer.